static/tests/samesite/form-post-blank.html

<!DOCTYPE html>
<script src="/tests/resources/testharness.js"></script>
<script src="/tests/resources/testharnessreport.js"></script>
<script src="/tests/resources/cookie-helper.js"></script>
<body>
<script>
  // Post to `/cookie/postToParent` on |origin|, returning a Promise which will
  // resolve with the cookie data it posts.
  function assertPopupReceivesCookie(origin, name, value) {
    return new Promise((resolve, reject) => {
      var f = document.createElement('form');
      f.action = origin + "/cookie/postToParent";
      f.target = "_blank";
      f.method = "POST";
      window.addEventListener("message", e => {
        // We know that there's only one window postMessaging here at a time (I hope).
        e.source.close();
        try {
          assert_cookie(origin, e.data, name, value, true);
          resolve("Popup received the cookie.");
        } catch (e) {
          reject(e);
        }
      });
      f.submit();
    });
  }

  // Invert the above results.
  function assertPopupDoesNotReceiveCookie(origin, name, value) {
    return new Promise((resolve, reject) => {
      assertPopupReceivesCookie(origin, name, value)
        .then(reject)
        .catch(_ => resolve("Popup did not receive the cookie."))
    })
  }

  function create_test(origin, target, extra, expected, title) {
    promise_test(t => {
      var COOKIE_VALUE = "" + Math.random();
      return create_cookie(origin, HTTP_COOKIE, COOKIE_VALUE, extra)
        .then(_ => {
          if (expected)
            return assertPopupReceivesCookie(target, HTTP_COOKIE, COOKIE_VALUE)
          else
            return assertPopupDoesNotReceiveCookie(target, HTTP_COOKIE, COOKIE_VALUE)
        });
    }, title);
  }

  // No redirect:
  create_test(ORIGIN, ORIGIN, "SameSite=Lax", true, "Same-host POSTs to '_blank' are laxly same-site");
  create_test(ORIGIN, ORIGIN, "SameSite=Strict", true, "Same-host POSTs to '_blank' are strictly same-site");
  create_test(SUBDOMAIN_ORIGIN, SUBDOMAIN_ORIGIN, "SameSite=Lax", true, "Subdomain POSTs to '_blank' are laxly same-site");
  create_test(SUBDOMAIN_ORIGIN, SUBDOMAIN_ORIGIN, "SameSite=Strict", true, "Subdomain POSTs to '_blank' are strictly same-site");
  create_test(CROSS_SITE_ORIGIN, CROSS_SITE_ORIGIN, "SameSite=Lax", false, "Cross-site POSTs to '_blank' are not laxly same-site");
  create_test(CROSS_SITE_ORIGIN, CROSS_SITE_ORIGIN, "SameSite=Strict", false, "Subdomain POSTs to '_blank' are not strictly same-site");

  // Redirect from {same-host,subdomain,cross-site} to same-host:
  create_test(ORIGIN, redirectTo(ORIGIN, ORIGIN), "SameSite=Lax", true, "Same-host redirecting to same-host POSTs to '_blank' are laxly same-site");
  create_test(ORIGIN, redirectTo(ORIGIN, ORIGIN), "SameSite=Strict", true, "Same-host redirecting to same-host POSTs to '_blank' are strictly same-site");
  create_test(ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, ORIGIN), "SameSite=Lax", true, "Subdomain redirecting to same-host POSTs to '_blank' are laxly same-site");
  create_test(ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, ORIGIN), "SameSite=Strict", true, "Subdomain redirecting to same-host POSTs to '_blank' are strictly same-site");
  create_test(ORIGIN, redirectTo(CROSS_SITE_ORIGIN, ORIGIN), "SameSite=Lax", true, "Cross-site redirecting to same-host POSTs to '_blank' are laxly same-site");
  create_test(ORIGIN, redirectTo(CROSS_SITE_ORIGIN, ORIGIN), "SameSite=Strict", true, "Cross-site redirecting to same-host POSTs to '_blank' are strictly same-site");

  // Redirect from {same-host,subdomain,cross-site} to subdomain:
  create_test(SUBDOMAIN_ORIGIN, redirectTo(ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Lax", true, "Same-host redirecting to subdomain POSTs to '_blank' are laxly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Strict", true, "Same-host redirecting to subdomain POSTs to '_blank' are strictly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Lax", true, "Subdomain redirecting to subdomain POSTs to '_blank' are laxly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Strict", true, "Subdomain redirecting to subdomain POSTs to '_blank' are strictly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(CROSS_SITE_ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Lax", true, "Cross-site redirecting to subdomain POSTs to '_blank' are laxly same-site");
  create_test(SUBDOMAIN_ORIGIN, redirectTo(CROSS_SITE_ORIGIN, SUBDOMAIN_ORIGIN), "SameSite=Strict", true, "Cross-site redirecting to subdomain POSTs to '_blank' are strictly same-site");

  // Redirect from {same-host,subdomain,cross-site} to cross-site:
  create_test(CROSS_SITE_ORIGIN, redirectTo(ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Lax", false, "Same-host redirecting to cross-site POSTs to '_blank' are not laxly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Strict", false, "Same-host redirecting to cross-site POSTs to '_blank' are not strictly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Lax", false, "Subdomain redirecting to cross-site POSTs to '_blank' are not laxly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(SUBDOMAIN_ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Strict", false, "Subdomain redirecting to cross-site POSTs to '_blank' are strictly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(CROSS_SITE_ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Lax", false, "Cross-site redirecting to cross-site POSTs to '_blank' are not laxly same-site");
  create_test(CROSS_SITE_ORIGIN, redirectTo(CROSS_SITE_ORIGIN, CROSS_SITE_ORIGIN), "SameSite=Strict", false, "Cross-site redirecting to cross-site POSTs to '_blank' are strictly same-site");
</script>
